move DetailText into StepList folder
