--- javascript/react/src/components/DetailText/index.ts
@@ -1,6 +1,0 @@
-/*!
- * Copyright (c) Microsoft. All rights reserved.
- * Licensed under the MIT license. See LICENSE file in the project.
- */
-export { DetailText } from './DetailText.js'
-export type { DetailTextProps } from './DetailText.types.js'
