+++ javascript/react/src/components/StepList/DetailText.styles.ts
