+++ javascript/react/src/components/StepList/DetailText.tsx
