+++ javascript/react/src/components/StepList/DetailText.types.ts
